Test interactor flow using argparse cli
