+++ course-management-tools/src/registrants/core/cli.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class CLI(ABC):
+    @abstractmethod
+    def get_input(self):
+        ...
